+++ src/components/basics/text-input-list/text-input-list.vue
@@ -1,155 +1,49 @@
 <template>
   <div :class="prefixCls">
-    <!-- 行号 -->
-    <div
-      v-for="(item, index) in maxLine"
-      :key="index"
-      :class="prefixCls + '-line'">
-      <div
-        :class="[
-          prefixCls + '-left-list',
-          { 'middle-style-li-active': middle.activeClass === index },
-          isHaveError(index) ? prefixCls + '-is-error' : ''
-        ]">
-        <span :class="[isHaveError(index) ? prefixCls + '-error-left' : '']">
-          {{ item }}
-        </span>
-      </div>
-
-      <div :class="prefixCls + '-right-list'">
-        <rd-text-emoj-input
-          :ref="`emojInput-${index}`"
-          :value="value[index] || ''"
-          :is-edit="middle.activeClass === index"
-          :class="[
-            { 'middle-style-li-active': middle.activeClass === index },
-            errors.includes(index) && prefixCls + '-is-error'
-          ]"
-          :transform-html2-text="transformHtml2Text"
-          :transform-text2-html="transformText2Html"
-          :calc-text-fn="calcTextFn"
-          :valid-fn="validFn"
-          :max-length="maxLength"
-          :min-length="minLength"
-          @on-keydown="handlerKeydown($event, index)"
-          @on-clear="handlerClear(index)"
-          @on-paste="handlerPaste($event, index)"
-          @input="val => handleEmitInput(val, index)"
-          @click.native="onClickEditorLine(index)"
-          @error="status => onError(status, index)" />
-        <div
-          v-if="(useEmoj || useEnter) && middle.activeClass === index"
-          :class="prefixCls + '-btn-wrap'"
-          @click.stop>
-          <Poptip
-            v-if="useEmoj"
-            v-model="showEmojPan"
-            transfer
-            :transfer-class-name="prefixCls + '-poptip'"
-            placement="bottom-end">
-            <img src="../../../images/text-input-list/add-emoji.png" />
-            <div
-              slot="content"
-              :class="prefixCls + '-panel-wrap'">
-              <img
-                v-for="(item, index) in emojiList"
-                :key="index"
-                :class="prefixCls + '-item-common'"
-                :src="item.url"
-                @click="insertFace(item)" />
-            </div>
-          </Poptip>
-
-          <img
-            v-if="useEnter"
-            src="../../../images/text-input-list/add-line-feed.png"
-            @click="insertEnter(index)" />
-        </div>
-      </div>
-    </div>
+    <virtual-list
+      ref="list"
+      class="virtual-list"
+      :data-key="'uid'"
+      :data-sources="getLine"
+      :extra-props="{ ...$attrs, middle, maxLine, errorList, value }"
+      :data-component="itemComponent"
+      v-on="$listeners"
+      @handlerKeydown="handlerKeydown"
+      @middle="middleChange"
+      @on-error="onInputError"
+      @itemClick="itemClick" />
   </div>
 </template>
-
 <script>
+import VirtualList from 'vue-virtual-scroll-list'
+import Item from './components/title-list'
 import { prefix } from '@src/config.js'
 const prefixCls = prefix + 'text-input-list'
-
-import rdTextEmojInput from '../text-emoj-input/text-emoj-input'
-import { getSplitReg } from '../../../util/text-emoj-input'
-
 export default {
   name: prefixCls,
   components: {
-    rdTextEmojInput
+    VirtualList,
+    Item
   },
   props: {
-    value: {
-      type: Array,
-      default: () => []
-    },
-    // 文本字符最小长度
-    minLength: {
-      type: Number,
-      default: 6
-    },
-    // 文本字符最大长度
-    maxLength: {
-      type: Number,
-      default: 30
-    },
     // 最大行数
     maxLine: {
       type: Number,
       default: 10
     },
-    // 最大换行数
-    maxEnter: {
-      type: Number,
-      default: 3
-    },
-    // 最大表情添加数
-    maxEmoji: {
-      type: Number,
-      default: 4
-    },
-    // 可以使用表情
-    useEmoj: {
-      type: Boolean,
-      default: true
-    },
-    // 可以使用换行
-    useEnter: {
-      type: Boolean,
-      default: true
-    },
-    // emoji表情列表
-    emojiList: {
+    value: {
       type: Array,
       default: () => []
-    },
-    // 文本计算方法
-    calcTextFn: {
-      type: Function,
-      default: text => {
-        // 默认每个文字算2个字符
-        const copyText = text.replaceAll(/[^\x00-\xff]/g, '**')
-        return copyText.length
-      }
-    },
-    // 错误校验方法
-    propsValidFn: {
-      type: Function
     }
   },
   data() {
     return {
       prefixCls,
+      itemChild: null,
+      itemComponent: Item,
       middle: {
-        defaultInput: '',
         preActiveClass: null,
         activeClass: null,
-        faceIcon: '',
-        addLineFeedIcon: require('../../../images/text-input-list/line-feed.png'),
         middleData: [
           {
             id: 1,
@@ -157,22 +51,19 @@ export default {
           }
         ]
       },
-      showEmojPan: false,
-      isError: false,
-      errors: []
+      errorList: []
     }
   },
   computed: {
-    curEmojInput() {
-      return (
-        this.$refs[`emojInput-${this.middle.activeClass}`] &&
-        this.$refs[`emojInput-${this.middle.activeClass}`][0]
-      )
-    },
-    isHaveError() {
-      return function (index) {
-        return this.errors.includes(index)
-      }
+    getLine() {
+      return Array(this.maxLine)
+        .fill('')
+        .map((item, idx) => ({
+          uid: `key_${idx}`,
+          index: idx,
+          item: idx + 1,
+          maxLine: this.maxLine
+        }))
     }
   },
   watch: {
@@ -181,162 +72,40 @@ export default {
     },
     value(cur) {
       if (cur.length === 0 && this.middle.activeClass !== null) {
-        this.curEmojInput.formatValue()
+        this.errorList = []
+        this.itemChild = null
       }
     }
   },
   methods: {
-    handlerKeydown({ keyDownEvent, disableInputFn }, index) {
-      // 回车换行
-      if (keyDownEvent.keyCode === 13) {
-        disableInputFn()
-        const curIndex = index + 1
-        if (curIndex >= this.maxLine) {
-          return
-        }
-        this.middle.activeClass = curIndex
-        this.$refs[`emojInput-${index}`][0].blur()
-        this.$nextTick(() => {
-          this.$refs[`emojInput-${curIndex}`][0].focus('end')
-        })
-      }
+    middleChange(middle) {
+      this.middle = middle
     },
-    handlerClear(index) {
-      const copyValue = JSON.parse(JSON.stringify(this.value))
-      copyValue.splice(index, 1)
-      this.$emit('input', copyValue)
-      this.$nextTick(() => {
-        this.curEmojInput.formatValue()
-        this.curEmojInput.focus('end')
-      })
+    onInputError(index, error) {
+      this.$set(this.errorList, index, error)
+      this.$emit('on-error', index, error)
     },
-    transformHtml2Text(html) {
-      return html
-    },
-    transformText2Html(text) {
-      return text
-    },
-    getFaceHtml(icon, type, name) {
-      const str = `<img style="pointer-events: none; margin-left: 4px; vertical-align: middle; " src="${icon}" draggable="false" width="16" height="16" data-type="${type}" data-name="${name}">`
-      return str
-    },
-    insertEnter() {
-      // 如果输入框内不存在内容，不允许点击换行符
-      if (!this.curEmojInput.getValue()) {
-        this.$Message.error('请先输入文本内容')
-        return
-      }
-      if (this.curEmojInput.getEnters() >= this.maxEnter) {
-        this.$Message.error(`最多能插入${this.maxEnter + 1}个换行`)
-        return
-      }
-      let html = `${this.getFaceHtml(this.middle.addLineFeedIcon, 'enter', '[回车]')}<br>&nbsp;`
-      this.curEmojInput.insertHtmlMark(html)
-    },
-    insertFace(val) {
-      if (this.curEmojInput.getEmojiNum() >= this.maxEmoji) {
-        this.$Message.warning(`建议不超过${this.maxEmoji}个表情包`)
-        return
-      }
-      this.faceIcon = val.url
-      let html = this.getFaceHtml(this.faceIcon, 'emoj', val.value)
-      this.curEmojInput.insertHtmlMark(html)
-    },
-    // 插入文本
-    insertText(text) {
-      let index = 0
-      if (this.middle.activeClass || this.middle.activeClass === 0) {
-        index = this.middle.activeClass
-      } else if (this.middle.preActiveClass) {
-        index = this.this.middle.preActiveClass
-      }
-      const com = this.$refs[`emojInput-${index}`][0]
-      if (this.middle.preActiveClass === null && this.middle.activeClass === null) {
-        com.$el.click()
-        this.$nextTick(() => {
-          com.insertText(text)
-        })
-        return
-      }
-      com.insertText(text)
-    },
-    // 向外抛出 input 事件，改变绑定数据
-    handleEmitInput(value, index) {
-      if (this.value[index] === value) {
-        return
-      }
-      const copyValue = JSON.parse(JSON.stringify(this.value))
-      copyValue[index] = value || ''
-      this.$emit('input', copyValue)
-    },
-    // 粘贴
-    handlerPaste(event, index) {
-      let itemList = event.clipboardData.items
-      for (let i = 0; i < itemList.length; i++) {
-        let item = itemList[i]
-        if (item.kind === 'string' && item.type.match('text/plain')) {
-          item.getAsString(str => {
-            let splitReg = getSplitReg(JSON.stringify(str).replace(/"/g, ''))
-            let arr = JSON.stringify(str)
-              .replace(/"/g, '')
-              .split(splitReg)
-              .filter(o => o)
-            const copyValue = JSON.parse(JSON.stringify(this.value))
-            arr.forEach((o, i) => {
-              if (i === 0) {
-                this.insertText(o || '')
-              } else if (index + i < this.maxLine) {
-                copyValue[index + i] = o
-              }
-            })
-            this.$emit('input', copyValue)
-          })
-        }
-      }
-    },
-    // 点击编辑行
-    onClickEditorLine(index) {
-      const el = this.$refs[`emojInput-${index}`][0]
-      this.middle.activeClass = index
-      this.$nextTick(() => {
-        el.focus()
-      })
-    },
-    // 错误类型
-    validFn(ln, value) {
-      let errors = []
-      if (ln && (ln > this.maxLength || ln < this.minLength)) {
-        errors.push('lengthError')
-      }
-      if (typeof this.propsValidFn === 'function') {
-        const allErrors = this.propsValidFn(value)
-        errors = [...errors, ...allErrors]
-      }
-      return errors
-    },
-    onError(errors, index) {
+    handlerKeydown(index) {
+      this.$refs.list.scrollToIndex(index - 5)
       setTimeout(() => {
-        const isHaveError = this.errors.indexOf(index)
-        if (isHaveError > -1) {
-          if (!errors.length) {
-            this.errors.splice(isHaveError, 1)
-          }
-        } else {
-          errors.length && this.errors.push(index)
-        }
-
-        this.$emit('on-error', index, errors)
+        const prevIndex = this.$children[0].$children.findIndex(item => item.index === index)
+        const emojInput = this.$children[0].$children[prevIndex].$children[0].$refs.emojInput
+        this.itemChild = this.$children[0].$children[prevIndex].$children[0]
+        emojInput.$el.click()
+        emojInput.focus('end')
       })
+    },
+    getItem() {
+      return this.itemChild || (this.$children[0] && this.$children[0].$children[0].$children[0])
     },
     getValue() {
-      let index = 0
-      if (this.middle.activeClass || this.middle.activeClass === 0) {
-        index = this.middle.activeClass
-      } else if (this.middle.preActiveClass) {
-        index = this.this.middle.preActiveClass
-      }
-      const com = this.$refs[`emojInput-${index}`][0]
-      return com.getValue()
+      return this.getItem().getValue()
+    },
+    insertText(val) {
+      this.getItem().insertText(val)
+    },
+    itemClick(child) {
+      this.itemChild = child
     }
   }
 }
